add step reordering

+++ routes/step.js
@@ -6,7 +6,8 @@ const express = require("express"),
 
 router.get("/", async (req, res) => {
 	try {
-		const foundSteps = await Step.find({task: req.params.taskId})
+		//Finds and sorts steps according to order
+		const foundSteps = await Step.find({task: req.params.taskId}).sort({"order": "ascending"}).exec();
 		res.json(foundSteps);
 	} catch(err) {
 		res.json(err);
@@ -16,7 +17,11 @@ router.get("/", async (req, res) => {
 router.post("/", async (req, res) => {
 	try {
 		const newStep = await Step.create({...req.body, task: req.params.taskId});
-		res.json(newStep)
+		const foundTask = await Task.findById(req.params.taskId);
+		//Adds step to task's references
+		await foundTask.steps.push(newStep._id);
+		await foundTask.save();
+		res.json(newStep);
 	} catch(err) {
 		res.json(err);
 	};
@@ -39,14 +44,32 @@ router.put("/:stepId", async (req, res) => {
 
 router.delete("/:stepId", async (req, res) => {
 	try {
-		if(!mongoose.Types.ObjectId.isValid(req.params.stepId)) {
-			return res.json({message: "Step does not exist."});
+		if(!mongoose.Types.ObjectId.isValid(req.params.stepId) || !mongoose.Types.ObjectId.isValid(req.params.taskId)) {
+			return res.json({message: "Step or task does not exist."});
 		};
+		const foundTask = await Task.findById(req.params.taskId);
+		//Removes step from task's references
+		await foundTask.steps.pull(req.params.stepId);
+		await foundTask.save();
 		const deletedStep = await Step.findByIdAndDelete(req.params.stepId);
-		if(!deletedStep) {
-			return res.json({message: "Step does not exist."});
+		if(!foundTask || !deletedStep) {
+			return res.json({message: "Step or task does not exist."});
 		}
 		res.json(req.params.stepId);
+	} catch(err) {
+		res.json(err);
+	};
+});
+
+router.post("/reorder", async (req, res) => {
+	try {
+		const foundSteps = await Step.find({task: req.params.taskId});
+		foundSteps.forEach(step => {
+			//Sets step's order to reordered step's index
+			step.order = req.body.map(el => el._id).indexOf(step._id.toString());
+			step.save();
+		});
+		res.json(foundSteps);
 	} catch(err) {
 		res.json(err);
 	};
